+++ src/main/java/utils/HandlebarsUtils.java
@@ -1,4 +1,4 @@
-package webserver;
+package utils;
 
 import com.github.jknack.handlebars.Handlebars;
 import com.github.jknack.handlebars.Template;
