+++ src/main/java/webserver/controller/UserCreateController.java
@@ -5,9 +5,9 @@ import java.net.URISyntaxException;
 
 import db.DataBase;
 import model.User;
-import webserver.HttpRequest;
-import webserver.HttpResponse;
-import webserver.Parameters;
+import webserver.http.HttpRequest;
+import webserver.http.HttpResponse;
+import webserver.http.Parameters;
 
 public class UserCreateController extends AbstractController {
