+++ src/main/java/webserver/http/HttpHeaders.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
