+++ src/main/java/webserver/http/HttpResponse.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
@@ -13,6 +13,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import utils.FileIoUtils;
+import utils.HandlebarsUtils;
 
 public class HttpResponse {
