+++ src/main/java/webserver/http/Protocol.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 import java.util.Objects;
