+++ src/test/java/utils/HandlebarsUtilsTest.java
@@ -9,8 +9,6 @@ import java.nio.charset.StandardCharsets;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import webserver.HandlebarsUtils;
-
 class HandlebarsUtilsTest {
 
     @DisplayName("view 이름으로 응답할 HTML 페이지를 찾는다.")
