+++ src/test/java/webserver/http/HttpHeadersTest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
@@ -8,6 +8,8 @@ import java.util.Map;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import webserver.http.HttpHeaders;
 
 class HttpHeadersTest {
