+++ src/test/java/webserver/http/HttpRequestTest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
@@ -8,6 +8,9 @@ import java.nio.charset.StandardCharsets;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import webserver.http.HttpMethod;
+import webserver.http.HttpRequest;
 
 class HttpRequestTest {
