+++ src/test/java/webserver/http/ParametersTest.java
@@ -1,10 +1,12 @@
-package webserver;
+package webserver.http;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import webserver.http.Parameters;
 
 class ParametersTest {
