+++ src/test/java/webserver/http/RequestPathTest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import exception.IllegalRequestPathException;
+import webserver.http.RequestPath;
 
 class RequestPathTest {
